lang-chain/requirements.txt
langchain
langchain-community
langchain-perplexity
python-dotenv
reactpy
fastapi
uvicorn[standard]
